Add files via upload
added ability to fetch the collected link from different websites. changed logic to look at the parent (product card) first, and then collect links and names, to ensure data consistency and to remove several bugs. Added ability to collect the description of item on its own page(the collected link). I wanted to collect only the ingredients, but it seems, that with my knowledge limitations and barbora and rimi website structure its too challenging to do that. Will stick to collecting all data and sending it to openAI next